refactor: Latest MVC Model Rewrite
