+++ asset/view/index.html
@@ -3,12 +3,12 @@
   <head>
     <meta charset="UTF-8" />
     <title>Chat</title>
-    <link rel="stylesheet" href="../../style/style_chat.css" />
-    <link rel="stylesheet" href="../../style/style_chat_folders.css" />
-    <link rel="stylesheet" href="../../style/style_chat_chatlist.css" />
-    <link rel="stylesheet" href="../../style/style_chat_dialog.css" />
-    <link rel="stylesheet" href="../../style/style_icons_and_fonts.css" />
-    <link rel="stylesheet" href="../../style/style_add_contacts.css" />
+    <link rel="stylesheet" href="../style/style_chat.css" />
+    <link rel="stylesheet" href="../style/style_chat_folders.css" />
+    <link rel="stylesheet" href="../style/style_chat_chatlist.css" />
+    <link rel="stylesheet" href="../style/style_chat_dialog.css" />
+    <link rel="stylesheet" href="../style/style_icons_and_fonts.css" />
+    <link rel="stylesheet" href="../style/style_add_contacts.css" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   </head>
 
@@ -67,7 +67,7 @@
           <div class="dialog__header">
             <div class="dialog__header-right">
               <img
-                src="../../image/user.png"
+                src="../image/user.png"
                 alt="profile"
                 class="message__img"
               />
@@ -229,6 +229,6 @@
       integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49"
       crossorigin="anonymous"
     ></script>
-    <script src="../../js/pageReaction.js"></script>
+    <script src="../js/pageReaction.js"></script>
   </body>
 </html>
